Add build and publish jobs
A new GH Workflow has been added to build and publish libEST to GH
Packages such that it can be used for CI without a COPR repo. The
libEST RPMs will be stored in an Alpine-based image to keep the
distribution small. The publishing will be done from v3.2.0-pki
branch since that branch has been designated to store PKI-specific
changes.

See also:
https://github.com/dogtagpki/libest/wiki